Bind actions to schedule search
